Refactored merged
