+++ frontend/src/App.js
@@ -26,10 +26,12 @@ function isLoggedIn(props) {
 function App (props) {
   return(
     <Router>
-      <header style={{fontFamily: 'san-serif',padding: '10px',
+{/*      <header style={{fontFamily: 'san-serif',padding: '10px',
       color: 'white',textAlign:'center',
       fontSize: '48px', backgroundColor: '#007AFF'}}>Airport Management System</header>
       <div style={{paddingBottom: '50px'}}></div>
+*/}
+      <div></div>
       <Routes>
         <Route path="/" element={!isLoggedIn(props) ? <Home/> : <Navigate to='/dashboard'><RoutesClass/> </Navigate>} />
         <Route exact path='/login' element={!isLoggedIn(props) ? <Login/> : <Navigate to='/dashboard'><RoutesClass/> </Navigate>} />
