+++ frontend/src/constants/sidebar-menu.js
@@ -1,13 +1,8 @@
 import DashboardIcon from '../assets/icons/dashboard.svg';
 
 import AddFlight from '../assets/images/AddFlights.jpeg';
-import ShippingIcon from '../assets/icons/shipping.svg';
 import ProductIcon from '../assets/icons/product.svg';
 import UserIcon from '../assets/icons/user.svg';
-import EnableGate from '../pages/EnableGate';
-import DisableGate from '../pages/DisableGate';
-import AssignGate from '../pages/AssignGate';
-import assignBaggageCarousel from '../pages/AssignBaggageCarousel';
 
 
 const sidebar_menu = [
@@ -42,18 +37,6 @@ const sidebar_menu = [
         icon: ProductIcon,
         path: '/dashboard/DisableGate',
         title: 'Disable Gate',
-    },
-    {
-        id: 6,
-        icon: ShippingIcon,
-        path: '/dashboard/AssignGate',
-        title: 'Assign Gate',
-    },
-    {
-        id: 7,
-        icon: UserIcon,
-        path: '/dashboard/assignBaggageCarousel',
-        title: 'Assign Baggage Carousel number'
     }
 ]
