+++ frontend/src/pages/styles.css
@@ -153,3 +153,33 @@ td span {
     font-weight: 700;
     border-color: #FFFFFF;
 }
+
+.dashbord-btn {
+    padding: 0.8em;
+    padding-left: 2em;
+    padding-right: 2em;
+    cursor: pointer;
+    color: #FFFFFF;
+    font-size: 0.9rem;
+    font-weight: 700;
+    border-radius: 8px;
+    text-transform: uppercase;
+    border: 1px solid #015C92;
+    background-color: #015C92;
+    margin: 0.2rem;
+}
+.dropdown {
+    display: flex;
+    flex-direction:row ;
+    gap:100px;
+    padding: 2em;
+    column-gap: 40px;
+    /*margin: 1em;
+    background-color: white;
+    border: none;
+    border-radius: 10px;
+    overflow: auto;
+    text-align: center;
+    align-items: center;
+    justify-content:space-around ;*/
+}
